Bug: rec_devs when MainRdevYrLast!=endyr
This fixes a bug where the number of forecast rec devs was asssumed equal to number of forecast years. Actually when the main rec devs period ends before the model end year these extra years are also included in the forecast rec devs vector. This is now accounted for by this code revision.